LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigation-traverseTo-same-document-preventDefault-multiple-windows-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: Navigation aborted


PASS navigation.traverseTo() - if a top window cancels the traversal, any iframes should not fire navigate
